unified namings, test for InvalidAvroObject
