--- src/AvroConvert/Exceptions/InvalidAvroHeaderException.cs
@@ -1,21 +1,0 @@
-﻿namespace AvroConvert.Exceptions
-{
-    using System;
-
-    public class InvalidAvroHeaderException : Exception
-    {
-        public InvalidAvroHeaderException()
-        {
-        }
-
-        public InvalidAvroHeaderException(string s)
-            : base(s)
-        {
-        }
-
-        public InvalidAvroHeaderException(string s, Exception inner)
-            : base(s, inner)
-        {
-        }
-    }
-}
