+++ src/AvroConvert/Exceptions/InvalidAvroObjectException.cs
@@ -1,0 +1,17 @@
+﻿namespace AvroConvert.Exceptions
+{
+    using System;
+
+    public class InvalidAvroObjectException : Exception
+    {
+        public InvalidAvroObjectException(string s)
+            : base(s)
+        {
+        }
+
+        public InvalidAvroObjectException(string s, Exception inner)
+            : base(s, inner)
+        {
+        }
+    }
+}
